requirements.txt
requests>=2.31
python-dotenv>=1.0
pandas>=2.1
numpy>=1.26
matplotlib>=3.8
seaborn>=0.13
plotly>=5.18
scipy>=1.12
scikit-learn>=1.4
xgboost>=2.0
streamlit>=1.30
